[2020-09-24] Event API_27 : Generator API DOCS code Refactoring
 - RestDocumentationResultHandler를 이용한 Docs snippet 생성 코드 분리 : AuthDocumentGenerator.java, EventDocumentGenerator.java
 - 미사용 설정 제거 : BaseControllerTest.java
 - docs 접근 요청 시 인증 처리 제외 경로 추가 : /favicon.ico, /error
