+++ src/main/asciidoc/index.adoc
@@ -104,14 +104,14 @@ include::{snippets}/errors/http-response.adoc[]
 
 `GET` 요청을 사용하여 인덱스에 접근할 수 있다.
 
-operation::index-api[snippets='response-body,http-response,links']
+operation::index[snippets='response-body,http-response,links']
 
 [[resources-auth]]
 === 인증 정보 요청
 
 `POST` 요청을 사용하여 인증 정보를 획득 합니다.
 
-operation::get-auth-info[snippets='request-headers,request-parameters,response-fields,response-headers']
+operation::auth-server-config-test/get-auth[snippets='request-headers,request-parameters,response-fields,response-headers']
 
 [[resources-events]]
 == 이벤트
@@ -123,26 +123,28 @@ operation::get-auth-info[snippets='request-headers,request-parameters,response-f
 
 `POST` 요청을 사용해서 신규 이벤트를 생성합니다.
 
-operation::create-event-api[snippets='curl-request,http-request,request-headers,request-body,request-fields,http-response,response-headers,response-fields,links']
-
-[[resources-events-get]]
-=== 이벤트 조회
-
-`Get` 요청을 사용해서 기존 이벤트 하나를 조회할 수 있다.
-
-operation::get-an-event-api[snippets='curl-request,http-request,request-headers,path-parameters,http-response,response-headers,response-body,response-fields,links']
-
-[[resources-events-update]]
-=== 이벤트 수정
-
-`PUT` 요청을 사용해서 기존 이벤트를 수정할 수 있다.
-
-operation::update-event-api[snippets='curl-request,http-request,request-headers,path-parameters,http-response,response-headers,response-body,response-fields,links']
+operation::event-controller-success-test/create-event-api[snippets='curl-request,http-request,request-headers,request-body,request-fields,http-response,response-headers,response-fields,links']
 
 [[resources-events-list]]
 === 이벤트 목록 조회
 
 `GET` 요청을 사용하여 조건내 이벤트 목록을 조회합니다.
 
-operation::get-event-list-api[snippets='response-fields,curl-request,http-response,links']
+operation::event-controller-success-test/get-event-list-api[snippets='response-fields,curl-request,http-response,links']
+
+[[resources-events-get]]
+=== 이벤트 조회
+
+`Get` 요청을 사용해서 기존 이벤트 하나를 조회할 수 있다.
+
+operation::event-controller-success-test/get-an-event-api[snippets='curl-request,http-request,request-headers,path-parameters,http-response,response-headers,response-body,response-fields,links']
+
+[[resources-events-update]]
+=== 이벤트 수정
+
+`PUT` 요청을 사용해서 기존 이벤트를 수정할 수 있다.
+
+operation::event-controller-success-test/update-event-api[snippets='curl-request,http-request,request-headers,path-parameters,http-response,response-headers,response-body,response-fields,links']
+
+
